Updated base-configuration, coreutils, file, libarchive, libevdev, libexpat and libffi
base-configuration: Modify PS1 to also show the hostname.
coreutils: Update to 8.32 and moved to the proper category.
file: Updated to 5.39.
libarchive: Update to 3.4.3 and moved to the proper category.
libevdev: Updated to 1.9.1 and moved to the proper category.
libexpat: Updated to 2.2.9 and moved to the proper category.
libffi: Updated to 3.3 and moved to the proper category.

meta: Sorted app-arch.yml and sys-apps.yml alphabetically

Signed-off-by: Dennisbonke <admin@dennisbonke.com>