scripts/vm-util: verify rootless USB passthrough
For setting up USB passthrough to not require running qemu with root,
the "uaccess" tag needs to be set for the device in udev. This commit
adds a check to verify that this has happened by resolving the device
path with `udevadm trigger`, and then checking attributes with
`udevadm info`.